fix: rename all files to CamelCase
