+++ youtube-client/src/app/app.module.ts
@@ -12,18 +12,18 @@ import { MatButtonModule } from '@angular/material/button';
 import AppRoutingModule from './app-routing.module';
 import AppComponent from './app.component';
 import HeaderComponent from './core/components/header/header.component';
-import SearchItemComponent from './core/components/search/search-item/search-item.component';
-import SearchResultComponent from './core/components/search/search-result/search-result.component';
-import SearchInputComponent from './core/components/search/search-input/search-input.component';
-import SearchSettingsComponent from './core/components/search/search-settings/search-settings.component';
+import SearchItemComponent from './core/components/search/searchItem/searchItem.component';
+import SearchResultComponent from './core/components/search/searchResult/searchResult.component';
+import SearchInputComponent from './core/components/search/searchInput/searchInput.component';
+import SearchSettingsComponent from './core/components/search/searchSettings/searchSettings.component';
 import LogoComponent from './core/components/header/logo/logo.component';
 import LoginComponent from './core/components/header/login/login.component';
 import FilterwordPipe from './shared/CustomPipes/filterword.pipe';
 import SortingvideoPipe from './shared/CustomPipes/sortingvideo.pipe';
-import MyColorDirective from './shared/Directives/my-color.directive';
+import MyColorDirective from './shared/Directives/myColor.directive';
 import ErrorComponent from './core/pages/error/error.component';
 import AuthComponent from './auth/pages/auth/auth.component';
-import CardDetailsComponent from './youtube/components/card-details/card-details.component';
+import CardDetailsComponent from './youtube/components/cardDetails/cardDetails.component';
 import MainpageComponent from './youtube/components/mainpage/mainpage.component';
 import HeaderInterceptor from './core/interceptor/header.interceptor';
