+++ youtube-client/src/app/core/components/search/searchInput/searchInput.component.scss
