+++ youtube-client/src/app/core/components/search/searchItem/searchItem.component.scss
