+++ youtube-client/src/app/core/components/search/searchResult/searchResult.component.scss
