+++ youtube-client/src/app/shared/CustomPipes/filterword.pipe.ts
@@ -1,5 +1,5 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import IItemVideoData from '../../core/components/search/search-item.model';
+import IItemVideoData from '../../core/components/search/searchItem.model';
 
 @Pipe({
   name: 'filterword',
