+++ youtube-client/src/app/youtube/components/cardDetails/cardDetails.component.ts
@@ -2,13 +2,13 @@ import { Component, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
 
-import IItemVideoData from 'src/app/core/components/search/search-item.model';
-import KeyAPIService from '../../services/youtuve-api.service';
+import IItemVideoData from 'src/app/core/components/search/searchItem.model';
+import KeyAPIService from '../../services/youtubeApi.service';
 
 @Component({
   selector: 'app-card-details',
-  templateUrl: './card-details.component.html',
-  styleUrls: ['./card-details.component.scss'],
+  templateUrl: './cardDetails.component.html',
+  styleUrls: ['./cardDetails.component.scss'],
 })
 export default class CardDetailsComponent implements OnInit {
   card: any;
